requirements.txt
PyAutoGUI~=0.9.53
PyGetWindow~=0.0.9
Pillow~=9.2.0